+++ packages/cli/src/zosfiles/-strings-/en.ts
@@ -586,6 +586,57 @@ export default {
         RESPONSETIMEOUT: "The maximum amount of time in seconds the z/OSMF Files TSO servlet should run before returning a response." +
         " Any request exceeding this amount of time will be terminated and return an error. Allowed values: 5 - 600"
     },
+    SEARCH: {
+        SUMMARY: "Search Files",
+        DESCRIPTION: "Search files for a search term.",
+        ACTIONS: {
+            DS: {
+                SUMMARY: "Search Data Sets and PDS Members",
+                DESCRIPTION: "Search all data sets and PDS members that match the data set name for a search term.",
+                POSITIONALS: {
+                    PATTERN: `The pattern to match data sets against. Also known as 'DSLEVEL'. The following special sequences can be ` +
+                    `used in the pattern:
+                    ${TextUtils.chalk.yellow("%")}: matches any single character
+                    ${TextUtils.chalk.yellow("*")}: matches any number of characters within a data set name qualifier ` +
+                    `(e.g. "ibmuser.j*.old" matches "ibmuser.jcl.old" but not "ibmuser.jcl.very.old")
+                    ${TextUtils.chalk.yellow("**")}: matches any number of characters within any number of data set name qualifiers ` +
+                    `(e.g. "ibmuser.**.old" matches both "ibmuser.jcl.old" and "ibmuser.jcl.very.old")
+                    However, the pattern cannot begin with any of these sequences.You can specify multiple patterns separated by commas, ` +
+                    `for example "ibmuser.**.cntl,ibmuser.**.jcl"`
+                },
+                OPTIONS: {
+                    CASESENSITIVE: "The search should be case sensitive",
+                    MAINFRAMESEARCH: "Perform a preliminary search on the mainframe, reducing network traffic",
+                    MAX_CONCURRENT_REQUESTS: "Specifies the maximum number of concurrent z/OSMF REST API requests to search files. " +
+                    "Increasing the value results in faster searches. " +
+                    "However, increasing the value increases resource consumption on z/OS and can be prone " +
+                    "to errors caused by making too many concurrent requests.",
+                    TIMEOUT: "The number of seconds to search before timing out"
+                },
+                EXAMPLES: {
+                    EX1: {
+                        DESCRIPTION: "Search all of IBMUSER's, data sets for 'ZOWE'",
+                        OPTIONS: "'IBMUSER.*' 'ZOWE'"
+                    },
+                    EX2: {
+                        DESCRIPTION: "Search all of IBMUSER's, data sets for 'ZOWE' in uppercase only",
+                        OPTIONS: "'IBMUSER.*' 'ZOWE' --case-sensitive"
+                    },
+                    EX3: {
+                        DESCRIPTION: "Search all of IBMUSER's, data sets for 'ZOWE', and time out in 1 minute",
+                        OPTIONS: "'IBMUSER.*' 'ZOWE' --timeout 60"
+                    },
+                    EX4: {
+                        DESCRIPTION: "Search all of IBMUSER's, data sets for 'ZOWE', and perform 8 parallel tasks",
+                        OPTIONS: "'IBMUSER.*' 'ZOWE' --max-concurrent-requests 8"
+                    }
+                }
+            }
+        },
+        OPTIONS: {
+            SEARCHSTRING: "The parameter to search for"
+        }
+    },
     UNMOUNT: {
         SUMMARY: "Unmount file systems",
         DESCRIPTION: "Unmount file systems, such as HFS, ZFS, and more. This disconnects you from USS file systems.",